# api/openapi.yaml
openapi: 3.1.0
info:
  title: Aurium daemon API
  version: 0.1.0-dev
  description: |
    D21: this API is the product boundary. The CLI, the embedded dashboard, a
    Tauri shell and a future cloud runtime are all clients of it, so anything
    only one client can do is a design mistake.

    This document is kept in step with the code by a test
    (internal/api/openapi_test.go) that fails when a route is registered
    without being described here, or described without being registered.
servers:
  - url: http://127.0.0.1:7770
    description: Loopback only. The daemon is per-user and never binds a public address.
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Host clients read ~/.aurium/token (mode 0600). Containers present a
        scoped token that may only act on their own container; scopes are
        context:*, ipc:*, task:*, snapshot:self and gateway:*.
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        root: { type: string }
        created_at: { type: string, format: date-time }
...
    Repository:
      type: object
      description: A git repository belonging to a project. A project may hold several (D22).
      properties:
        id: { type: string }
        project_id: { type: string }
        path: { type: string }
        base_branch: { type: string }
        remote: { type: string }
...
    Approval:
      type: object
      properties:
        id: { type: string }
        capability: { type: string }
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        expires_at: { type: string, format: date-time }
paths:
...
  /v1/events:
    get:
      summary: Server-sent event stream
      description: |
        Replays from `since` before the live stream resumes, so a reconnecting
        client neither misses nor duplicates events. Frames carry no `event:`
        field — the type is in the JSON payload, because a typed SSE event can
        only be received through a per-type listener and the set of Aurium
        event types is open.

        This route alone also accepts the token as a query parameter, because
        EventSource cannot set request headers.
      parameters:
        - { name: since, in: query, schema: { type: integer } }
        - { name: types, in: query, schema: { type: string } }
        - { name: container, in: query, schema: { type: string } }
        - { name: token, in: query, schema: { type: string } }
      responses:
        "200": { description: text/event-stream }
  /v1/projects:
    post:
      summary: Create a project
      description: |
        D22: a project is a set of repositories described by aurium.project.yaml.
        This route and `aurium init` call the same provisioning code, so neither
        can be the one that really works (D21).

        A member repository that cannot be attached fails the whole call. Half a
        project on disk is worse than a clean failure, because nothing cleans it
        up.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                root:
                  type: string
                  description: |
                    Where the descriptor is written. Defaults to
                    ~/.aurium/projects/<slug>, which keeps it out of any one
                    member repository — writing it into one would make that repo
                    quietly special.
                repos:
                  type: array
                  items:
                    type: object
                    properties:
                      path: { type: string }
                      base_branch: { type: string }
                context: { type: array, items: { type: string } }
                driver: { type: string }
                image: { type: string }
                agent: { type: string }
      responses:
        "201": { description: Created }
        "400": { description: A member is not a git repository, or the root is already a project }
        "403": { description: An agent may not create projects }
    get:
      summary: List projects
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects: { type: array, items: { $ref: "#/components/schemas/Project" } }
  /v1/projects/{project}:
    get:
      summary: One project, with its repositories
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { $ref: "#/components/schemas/Project" }
                  repositories: { type: array, items: { $ref: "#/components/schemas/Repository" } }
                  containers: { type: integer }
                  agents: { type: integer }
                  standalone:
                    type: boolean
                    description: True for a single repository registered by `aurium init` with no descriptor.
        "404": { description: Not found }
  /v1/projects/{project}/repos:
    post:
      summary: Attach a repository to a project
      description: |
        Idempotent within a project: re-attaching a member returns the existing
        row, because "add these three repos" where one is already a member is an
        ordinary request. A repository already in a *different* project is
        refused — a repository belongs to one project at a time.

        Host-only. An agent able to attach repositories could register anything
        on the host.
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: Any directory in the repository; resolved to its git toplevel.
                base_branch: { type: string }
                driver: { type: string }
                image: { type: string }
                agent: { type: string }
      responses:
        "201":
          description: Attached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Repository" }
        "400": { description: Not a git repository, or already in another project }
        "403": { description: An agent may not attach repositories }
  /v1/projects/{project}/agents:
    post:
      summary: Create an agent (and the container it needs)
      description: |
        "Add an agent" is the user's unit. A container — a worktree on a branch
        — is the mechanism, created here because an agent needs one, not because
        anybody asked for one.

        Host-only. An agent able to spawn agents could fan out without bound;
        delegation (§9.4) is the path for that and it has a depth cap.

        With one repository in the project the choice of where to put it is
        made; with several, `repo_id` is required rather than guessed, because
        guessing puts an agent in the wrong tree.
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                repo_id: { type: string }
                name:
                  type: string
                  description: What the rail shows. Defaults to "Agent N".
                branch:
                  type: string
                  description: Defaults to a slug of the name, made unique.
                adapter: { type: string, description: "Defaults to the repo's aurium.yaml." }
                model: { type: string }
                role: { type: string, enum: [primary, master, worker] }
                task: { type: string, description: Creates a task and attaches the container to it. }
                parent_container_id: { type: string, description: Stack on an existing container. }
                provider_account_id: { type: string }
      responses:
        "201": { description: Created }
        "400": { description: The branch exists, the repo is ambiguous, or the driver refused }
        "403": { description: An agent may not spawn agents }
    get:
      summary: Every agent in a project, as rail tiles
      description: |
        One query across every container, because the rail redraws on every
        event and a call per container would not survive that.
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents: { type: array, items: { $ref: "#/components/schemas/AgentTile" } }
  /v1/projects/{project}/containers:
    get:
      summary: List a project's containers
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
  /v1/projects/{project}/tasks:
    get:
      summary: List tasks
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
    post:
      summary: Create a task
      description: Requires the task:* scope.
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                parent_task_id: { type: string }
      responses:
        "201": { description: Created }
        "403": { description: The token lacks task:* }
  /v1/projects/{project}/tree:
    get:
      summary: The container forest with live sync status
      description: |
        Sync status is recomputed from git per request rather than read from
        the stored status, which can be seconds out of date.
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
  /v1/agents/{agent}:
    get:
      summary: One agent, its container and the account it is billed against
      parameters: [{ name: agent, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "403": { description: This token is scoped to a different container }
        "404": { description: Not found }
  /v1/agents/{agent}/messages:
    get:
      summary: An agent's transcript
      description: |
        Read-only, and deliberately not the inbox. At-least-once delivery marks
        a message delivered when the recipient is handed it, and a human reading
        over an agent's shoulder is not the recipient.
      parameters:
        - { name: agent, in: path, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200": { description: OK }
        "403": { description: This token is scoped to a different container }
  /v1/agents/{agent}/message:
    post:
      summary: Say something to an agent
      description: |
        Two things happen. The message is recorded as IPC, so it is in the
        agent's inbox whether or not it is looking; and for an interactive agent
        on a driver with tmux it is typed into the session, so an agent sitting
        at a prompt sees it now. The response says which, so the dashboard can
        report "queued" rather than implying it has been seen.
      parameters: [{ name: agent, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                type: { type: string, default: REQUEST }
                priority: { type: string, enum: [normal, high] }
      responses:
        "201": { description: Sent }
        "400": { description: Empty content or an unknown message type }
        "403": { description: This token is scoped to a different container }
  /v1/containers/{container}:
    get:
      summary: Get a container
      description: |
        A container-scoped token may only address its own container; any other
        returns 403.
      parameters: [{ name: container, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "403": { description: This token is scoped to a different container }
        "404": { description: Not found }
  /v1/containers/{container}/agents:
    get:
      summary: List a container's agents
      parameters: [{ name: container, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
  /v1/containers/{container}/snapshots:
    get:
      summary: List a container's snapshots
      parameters: [{ name: container, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
  /v1/containers/{container}/snapshot:
    post:
      summary: Take a snapshot
      description: Requires snapshot:self, which does not permit snapshotting another container.
      parameters: [{ name: container, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { label: { type: string } }
      responses:
        "201": { description: Created }
  /v1/containers/{container}/sync:
    post:
      summary: Rebase a container onto its parent's tip
      description: |
        A conflict is a normal outcome and returns 200 with result "conflict"
        and the conflicted files, not an error status.
      parameters: [{ name: container, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: The outcome, including conflicts }
  /v1/tasks/{task}:
    patch:
      summary: Transition a task
      parameters: [{ name: task, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { status: { type: string } }
      responses:
        "200": { description: OK }
        "400": { description: Unknown status }
  /v1/approvals:
    get:
      summary: List approvals
      description: Defaults to pending, which is the inbox a human actually wants.
      parameters: [{ name: status, in: query, schema: { type: string } }]
      responses:
        "200": { description: OK }
  /v1/approvals/{approval}/decide:
    post:
      summary: Approve or reject a held call
      description: |
        Approving EXECUTES the held upstream call in the daemon, which is why
        this cannot be done in the CLI process: only the daemon holds the
        connected upstreams and their credentials (D17). The result is
        delivered to the agent as a RESPONSE message, not returned here.
      parameters: [{ name: approval, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                by: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Approval" }
  /v1/providers:
    get:
      summary: Connected provider accounts
      description: |
        Returns keyring references and never credentials. There is deliberately
        no route that resolves a reference back to a secret (D17, D24).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts: { type: array, items: { $ref: "#/components/schemas/ProviderAccount" } }
        "403": { description: Host-only }
    post:
      summary: Connect a provider account
      description: |
        Three ways in, and exactly one of them stores nothing. `secret` pastes a
        key or a token; `use_host_env` takes the value this host already exports;
        `use_cli_login` records that the provider's own CLI is logged in here and
        holds no copy at all.

        D25: there is no OAuth client Aurium could legitimately drive for a Claude
        or ChatGPT subscription, so a seat is reached through the provider CLI's
        own login (`claude setup-token`, `codex login`) and the token it prints.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [provider, auth_kind]
              properties:
                provider: { type: string, enum: [anthropic, openai] }
                auth_kind: { type: string, enum: [api_key, subscription] }
                label: { type: string }
                secret:
                  type: string
                  description: Never echoed back, never logged, never stored outside the keyring.
                use_host_env: { type: boolean }
                use_cli_login: { type: boolean }
      responses:
        "201":
          description: Connected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderAccount" }
        "400": { description: No credential, an unset variable, or a key for the other provider }
        "409": { description: An account with that label already exists for this provider }
        "403": { description: Host-only }
  /v1/providers/detect:
    get:
      summary: What this host already offers
      description: |
        Names the environment variables that are set and the CLI login files that
        exist. It reports whether a credential exists, never what it is.
      responses:
        "200": { description: OK }
        "403": { description: Host-only }
  /v1/providers/{account}:
    delete:
      summary: Disconnect an account
      description: |
        Deletes the credential as well as the row. Agents running on it keep
        running — the credential is already in their environment — and their link
        is cleared rather than the agents being deleted with it.
      parameters: [{ name: account, in: path, required: true, schema: { type: string } }]
      responses:
        "204": { description: Disconnected }
        "404": { description: Not found }
        "403": { description: Host-only }
  /v1/github:
    get:
      summary: Whether Aurium can see your GitHub
      description: |
        D28: the token comes from `gh auth token` when `gh` is logged in, and
        from the keyring otherwise. Storing a second copy of a credential the
        machine already holds creates a second thing to revoke and a second
        thing to go stale.

        The token is checked rather than assumed — a stale one that still looks
        like a token would otherwise read as connected until the first listing
        failed for no visible reason. Host-only: an agent reaches GitHub through
        the MCP gateway under a grant (§10).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected: { type: boolean }
                  source: { type: string, enum: [gh_cli, stored] }
                  login: { type: string }
                  name: { type: string }
                  error: { type: string, description: A token that exists but does not work. }
                  cli_available: { type: boolean }
        "403": { description: Host-only }
  /v1/github/repos:
    get:
      summary: The repositories this token can see
      description: Most recently pushed first, which is nearly always the order somebody wants.
      parameters: [{ name: limit, in: query, schema: { type: integer, default: 100 } }]
      responses:
        "200": { description: OK }
        "400": { description: No GitHub connection, or GitHub refused the token }
        "502": { description: GitHub could not be reached }
        "403": { description: Host-only }
  /v1/projects/{project}/github/clone:
    post:
      summary: Clone a GitHub repository into a project and attach it
      description: |
        One call because it is one intent: a repository on disk that no project
        knows about is not what anybody asked for.

        The destination is confined to the project root. Without that, `dest` is
        an arbitrary filesystem path a request can name, and this route would
        clone anything anywhere as the user — far more authority than "add a
        repository to this project" asks for.

        The token reaches git through GIT_ASKPASS, never in the URL: a URL with
        a token in it ends up in `git remote -v`, in .git/config, and in every
        log that records the command.
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [clone_url]
              properties:
                clone_url: { type: string }
                full_name: { type: string, description: owner/repo, used to name the directory. }
                dest: { type: string, description: Defaults to <project root>/<repo name>. }
                base_branch: { type: string }
                driver: { type: string }
                image: { type: string }
                agent: { type: string }
      responses:
        "201": { description: Cloned and attached }
        "400": { description: No connection, a non-empty destination, or a path outside the project }
        "403": { description: Host-only }
  /v1/projects/{project}/github/tools:
    post:
      summary: Give this project's agents github_* capabilities
      description: |
        The other GitHub: the MCP gateway's (§10). Agents never touch the token
        — the daemon holds it, spawns the upstream server and filters which
        tools each container may call. Ungranted tools are not listed at all, so
        an agent cannot try one it does not hold.

        The token is borrowed from `gh auth token` at connect time and never
        written down. It is re-read on every daemon start, because `gh` rotates
        it and a keyring copy would go stale silently.
      parameters: [{ name: project, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: Connected, with the capability count that was registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  integration: { type: string }
                  status: { type: string }
                  last_error: { type: string }
                  capabilities: { type: integer }
        "400": { description: "`gh` is not logged in on this machine" }
        "502": { description: The upstream MCP server could not be started }
        "403": { description: Host-only }
  /v1/usage:
    get:
      summary: What the fleet has spent
      description: |
        Grouped by provider, agent, model and project in one response, because
        four round trips to draw one screen is four chances for the parts to
        disagree. The payload states what the figures cover: list prices only,
        and only calls a provider actually reported.
      parameters:
        - { name: window, in: query, schema: { type: string, default: 24h } }
        - { name: project, in: query, schema: { type: string } }
        - { name: agent, in: query, schema: { type: string } }
        - { name: provider, in: query, schema: { type: string } }
      responses:
        "200": { description: OK }
        "400": { description: The window is not a duration, or is longer than a year }
        "403": { description: Host-only }
  /v1/usage/series:
    get:
      summary: Usage bucketed over time
      description: Bucket width is chosen to give roughly 48 points whatever the window.
      parameters:
        - { name: window, in: query, schema: { type: string, default: 24h } }
        - { name: project, in: query, schema: { type: string } }
      responses:
        "200": { description: OK }
        "403": { description: Host-only }
  /v1/heartbeat:
    get:
      summary: Live vitals for the dashboard's centre pane
      description: |
        Exists so the pane is alive before the event stream has warmed up; the
        stream animates it afterwards. Rates rather than lifetime totals: a
        counter that only climbs says nothing about now.
      parameters: [{ name: project, in: query, schema: { type: string } }]
      responses:
        "200": { description: OK }
  /mcp:
    post:
      summary: MCP endpoint reached by the in-container aurium-mcp shim
      description: |
        The caller is derived entirely from the bearer token, never from the
        request body, and the agent's role is read from the database — an
        agent cannot promote itself by claiming a role in the payload.

        JSON-RPC errors travel inside a 200 so the shim hands the agent an
        error it can act on rather than reporting a broken transport. A
        notification returns 202 with no body.
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: A JSON-RPC response, possibly carrying an error object }
        "202": { description: A notification was accepted }
        "403": { description: The token lacks gateway:* }
...